Move management commands from mozorg to base
