+++ bedrock/base/management/commands/update_product_details_files.py
